requirements.txt
Flask==0.10.1
python-dotenv==0.9.1
requests==2.19.1
urllib3==1.23
Werkzeug==0.14.1
